move types to .protocol
